revert ObjIdx to uint8; fix bigblk heuristic for targetObjSize=0
Revert ObjIdx from uint16 back to uint8. Production merges are resource-
throttled to ~2-4 GB input, producing at most ~16-32 output objects
(128 MB each). Exceeding 255 is not realistic.

The overflow guard (error if objCnt >= NoTransfer) is retained to prevent
silent wrapping.

Also fix the bigblk heuristic: when targetObjSize=0 (block-count mode),
the check 'buffer.Size() > 0*3/2' was always true, creating a new object
every ~100 rows. Skip bigblk check when targetObjSize is 0.

+++ pkg/vm/engine/tae/rpc/handle_debug.go
@@ -1048,7 +1048,7 @@ func marshalTransferMaps(
 				for i := range bat.RowCount() {
 					srcBlk := vector.GetFixedAtNoTypeCheck[int32](bat.Vecs[0], i)
 					srcRow := vector.GetFixedAtNoTypeCheck[uint32](bat.Vecs[1], i)
-					destObj := vector.GetFixedAtNoTypeCheck[uint16](bat.Vecs[2], i)
+					destObj := vector.GetFixedAtNoTypeCheck[uint8](bat.Vecs[2], i)
 					destBlk := vector.GetFixedAtNoTypeCheck[uint16](bat.Vecs[3], i)
 					destRow := vector.GetFixedAtNoTypeCheck[uint32](bat.Vecs[4], i)
 
@@ -1085,7 +1085,7 @@ func marshalTransferMaps(
 			}
 			for r, pos := range m.M {
 				tm[uint32(r)] = api.TransferDestPos{
-					ObjIdx: uint16(pos.ObjIdx),
+					ObjIdx: uint8(pos.ObjIdx),
 					BlkIdx: uint16(pos.BlkIdx),
 					RowIdx: uint32(pos.RowIdx),
 				}
